Refactor assistants endpoints into modular management and tools routers
